requirements.txt
Package                       Version
----------------------------- ----------
asgiref                       3.7.2
certifi                       2023.11.17
cffi                          1.16.0
charset-normalizer            3.3.2
cryptography                  42.0.2
defusedxml                    0.8.0rc2
Django                        5.0.1
django-rest                   0.8.7
django-rest-framework         0.1.0
django-templated-mail         1.1.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
djoser                        2.2.2
docker                        7.0.0
drf-yasg                      1.21.7
idna                          3.6
inflection                    0.5.1
oauthlib                      3.2.2
packaging                     23.2
pip                           23.3.2
postgres                      4.0
psycopg2                      2.9.9
psycopg2-binary               2.9.9
psycopg2-pool                 1.1
pycparser                     2.21
PyJWT                         2.8.0
python3-openid                3.2.0
pytz                          2023.4
PyYAML                        6.0.1
requests                      2.31.0
requests-oauthlib             1.3.1
setuptools                    65.5.0
six                           1.16.0
social-auth-app-django        5.4.0
social-auth-core              4.5.2
sqlparse                      0.4.4
uritemplate                   4.1.1
urllib3                       2.1.0
